# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fnashha Home Services Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/customer:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/register/technician:
    post:
      operationId: registerTechnician
      tags: [auth]
      summary: Register a new technician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /commissions:
    get:
      operationId: listCommissions
      tags: [commissions]
      summary: List commission settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commission"
    post:
      operationId: createCommission
      tags: [commissions]
      summary: Create commission rule (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
  /commissions/{id}:
    patch:
      operationId: updateCommission
      tags: [commissions]
      summary: Update commission rule (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"

  # ─── SUPPORT ─────────────────────────────────────────────────────────────────
...
  /cms/settings:
    get:
      operationId: getCmsSettings
      tags: [cms]
      summary: Get CMS settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CmsSettings"
    patch:
      operationId: updateCmsSettings
      tags: [cms]
      summary: Update CMS settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmsSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CmsSettings"

  # ─── ANALYTICS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CustomerRegistration:
      type: object
      required: [fullName, mobile, password]
      properties:
        fullName:
          type: string
        mobile:
          type: string
        password:
          type: string
    TechnicianRegistration:
      type: object
      required: [fullName, mobile, password, nationalId, serviceIds, governorateIds, areaIds, primaryAreaId]
      properties:
        fullName:
          type: string
        mobile:
          type: string
        password:
          type: string
        nationalId:
          type: string
        personalPhoto:
          type: ["string", "null"]
        nationalIdFront:
          type: ["string", "null"]
        nationalIdBack:
          type: ["string", "null"]
        serviceIds:
          type: array
          items:
            type: integer
        governorateIds:
          type: array
          items:
            type: integer
        areaIds:
          type: array
          items:
            type: integer
        primaryAreaId:
          type: integer
        acceptTerms:
          type: boolean
    LoginInput:
      type: object
      required: [mobile, password]
      properties:
        mobile:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, technician, admin, super_admin]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── USER ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, fullName, mobile, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, technician, admin, super_admin]
        status:
          type: string
          enum: [active, pending, suspended, banned, rejected]
        profileImage:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
        technicianProfile:
          $ref: "#/components/schemas/TechnicianProfile"
...
    TechnicianProfile:
      type: object
      required: [id, userId, nationalId, approvalStatus]
      properties:
        id:
          type: integer
        userId:
          type: integer
        nationalId:
          type: string
        personalPhoto:
          type: ["string", "null"]
        nationalIdFront:
          type: ["string", "null"]
        nationalIdBack:
          type: ["string", "null"]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        pointsBalance:
          type: integer
        averageRating:
          type: number
        reviewCount:
          type: integer
        completedJobs:
          type: integer
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        areas:
          type: array
          items:
            $ref: "#/components/schemas/Area"
        primaryArea:
          $ref: "#/components/schemas/Area"
...
    Service:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
...
    Governorate:
      type: object
      required: [id, name, nameAr, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
...
    Area:
      type: object
      required: [id, name, nameAr, governorateId, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        governorate:
          $ref: "#/components/schemas/Governorate"
        isActive:
          type: boolean
...
    Commission:
      type: object
      required: [id, type, value, createdAt]
      properties:
        id:
          type: integer
        serviceId:
          type: ["integer", "null"]
        service:
          $ref: "#/components/schemas/Service"
        areaId:
          type: ["integer", "null"]
        area:
          $ref: "#/components/schemas/Area"
        type:
          type: string
          enum: [fixed, percentage]
        value:
          type: number
        createdAt:
          type: string
    CommissionInput:
      type: object
      required: [type, value]
      properties:
        serviceId:
          type: integer
        areaId:
          type: integer
        type:
          type: string
          enum: [fixed, percentage]
        value:
          type: number
    CommissionUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [fixed, percentage]
        value:
          type: number

    # ─── SUPPORT ───────────────────────────────────────────────────────────────
...
    CmsSettings:
      type: object
      properties:
        siteName:
          type: string
        siteNameAr:
          type: string
        logoUrl:
          type: ["string", "null"]
        faviconUrl:
          type: ["string", "null"]
        heroTitle:
          type: string
        heroTitleAr:
          type: string
        heroSubtitle:
          type: string
        heroSubtitleAr:
          type: string
        aboutUs:
          type: string
        aboutUsAr:
          type: string
        termsConditions:
          type: string
        termsConditionsAr:
          type: string
        privacyPolicy:
          type: string
        privacyPolicyAr:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        contactAddress:
          type: string
        facebookUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
    CmsSettingsUpdate:
      type: object
      properties:
        siteName:
          type: string
        siteNameAr:
          type: string
        heroTitle:
          type: string
        heroTitleAr:
          type: string
        heroSubtitle:
          type: string
        heroSubtitleAr:
          type: string
        aboutUs:
          type: string
        aboutUsAr:
          type: string
        termsConditions:
          type: string
        termsConditionsAr:
          type: string
        privacyPolicy:
          type: string
        privacyPolicyAr:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        contactAddress:
          type: string
        facebookUrl:
          type: string
        twitterUrl:
          type: string
        instagramUrl:
          type: string
        whatsappNumber:
          type: string

    # ─── ANALYTICS ─────────────────────────────────────────────────────────────
...